Update README.md
This project tries to find out the job satisfaction level of individuals based on various parameters taking into consideration Job satisfaction is defined as the level of contentment employees feel with their job. This goes beyond their daily duties to cover satisfaction with team members and the impact of their job on employees' personal lives. This project provides insights on various factors which decide job satisfaction level of an individual. The insights will help an individual to focus/improve his/her job satisfaction level. For this Project we did Primary data collection based on prepare Surveys. Did all the EDA part. And finally reach on conclusions based as per the statistics report showing.